Refactor body of HomePage. - Extract PageView into home_app_body. - Extract views to their own classes.
