+++ lib/pages/home/home_page.dart
@@ -1,16 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:school_erp/features/auth/auth_repository/schemas/user.dart';
 import 'package:school_erp/pages/home/utils/app_bar/home_app_bar.dart';
+import 'package:school_erp/pages/home/utils/app_body/home_app_body.dart';
 import 'package:school_erp/pages/home/utils/bottom_nav/bottom_nav_destinations.dart';
 import 'package:school_erp/pages/home/utils/bottom_nav/bottom_navigation.dart';
-import 'package:school_erp/pages/home/widgets/dashboard_header.dart';
-import 'package:school_erp/pages/home/widgets/features.dart';
-import 'package:school_erp/pages/home/widgets/feeds.dart';
 // import 'package:school_erp/pages/home/widgets/message.dart';
-import 'package:school_erp/pages/home/widgets/settings.dart';
-import 'package:school_erp/pages/home/widgets/time_record_card.dart';
 import 'package:school_erp/theme/colors.dart';
-import 'package:school_erp/theme/text_styles.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage(this.user, {super.key});
@@ -42,95 +37,25 @@ class _HomePageState extends State<HomePage> {
     );
   }
 
+  void _onPageChangedFn(index) {
+    setState(() {
+      currentPageIndex = index;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: HomeAppBar(user: widget.user),
+      body: HomeAppBody(
+          user: widget.user,
+          pageController: _pageController,
+          onPageChangedFn: _onPageChangedFn),
       bottomNavigationBar: BottomNavigation(
           tapFn: _navOnTap,
           currentPageIndex: currentPageIndex,
           destinations: BottomNavDestinations()),
       backgroundColor: AppColors.primaryColor,
-      body: PageView(
-        controller: _pageController,
-        onPageChanged: (index) {
-          setState(() {
-            currentPageIndex = index;
-          });
-        },
-        children: [
-          HomeWidget(user: widget.user),
-          FeedsWidget(user: widget.user),
-          ComingSoonWidget(),
-          //MessageWidget(user: widget.user),
-          SettingWidget(user: widget.user),
-        ],
-      ),
-    );
-  }
-}
-
-class HomeWidget extends StatelessWidget {
-  const HomeWidget({super.key, required this.user});
-
-  final AuthenticatedUser user;
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      child: Column(
-        mainAxisSize: MainAxisSize.max,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              children: [
-                DashboardHeader(user: user),
-                const SizedBox(height: 16),
-                const TimeRecordWidget(),
-              ],
-            ),
-          ),
-          Expanded(child: Features(user: user)),
-        ],
-      ),
-    );
-  }
-}
-
-class ComingSoonWidget extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          const Text(
-            'Chat rooms are under construction!!!',
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              fontSize: 24,
-              fontWeight: FontWeight.bold,
-              color: AppColors.whiteColor,
-            ),
-          ),
-          const SizedBox(height: 16),
-          Image.asset(
-            'assets/images/under_construction.webp',
-          ),
-          const SizedBox(height: 16),
-          const Text(
-            "Thank you for your patience! \n Stay tuned for what's coming soon!",
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.bold,
-              color: AppColors.whiteColor,
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
